Refactor: Move logic to data source
