--- data/src/main/java/com/android254/data/repos/remote/RemoteFeedDataSource.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2023 DroidconKE
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.android254.data.repos.remote
-
-import com.android254.data.network.apis.FeedApi
-import com.android254.data.network.models.responses.FeedDTO
-import com.android254.domain.models.DataResult
-import javax.inject.Inject
-
-interface RemoteFeedDataSource {
-
-    suspend fun fetchFeed(): DataResult<List<FeedDTO>>
-}
-
-class RemoteFeedDataSourceImpl @Inject constructor(
-    private val api: FeedApi
-) : RemoteFeedDataSource {
-
-    override suspend fun fetchFeed(): DataResult<List<FeedDTO>> =
-        api.fetchFeed()
-}
